backend/pyproject.toml
[project]
name = "MewAI-Backend"
version = "0.2.1" # Versión incrementada
description = "Backend for MewAI using crewAI and FastAPI"
authors = [{ name = "David Silvera", email = "ingenieria.d.s.g@hotmail.com" }]
requires-python = ">=3.10,<=3.13"
dependencies = [
    "crewai[tools]>=0.86.0,<1.0.0",
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn[standard]>=0.29.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "PyYAML>=6.0,<7.0.0",
    "gradio_client>=0.16.0,<1.0.0",
    "aiofiles>=23.2.1,<24.0.0",
    # --- DEPENDENCIA ACTUALIZADA ---
    # langchain-core y langchain-google-genai ya no son estrictamente necesarias aquí
    # crewai.LLM usa LiteLLM bajo el capó. Incluirlo explícitamente es bueno.
    "litellm>=1.35.0,<2.0.0" # O versión compatible que use crewai
]

[project.scripts]
# Mantén estos si los usas, si no, puedes eliminarlos
# mininos = "mininos.src.main:main"
# run_crew = "mininos.src.main:main"
# train = "mininos.src.main:train"
# replay = "mininos.src.main:replay"
# test = "mininos.src.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]